+++ software_stm32/include/compile_time.h
@@ -79,5 +79,3 @@
  
 #endif /* COMPILE_TIME_H_ */
 
-
-