+++ software_stm32/ChangeLog.md
@@ -11,6 +11,10 @@ All notable changes to this project will be documented in this file.
 
 ### Removed
 
+## [1.3.7]
+### Fixed
+- Config :      fix bug valve 12 set tIdx1 and tIdx2 was not saved
+
 ## [1.3.6]
 ### Changed
 - rework motor end detection (filtering of motor current, no soft PWM for motor turn on anymore)